Removed float and num instantiations of mat-vec-3 from package
The chicken compiler seems to be generating erroneus import files for these.
The instantiations will compile and install without error; however, any attempt
to import these modules (e.g. fp-mat-vec-3) will fail. It looks as though the
symbol M does not get bound to the module that the functor is supposed to map.

Instantiating a functor in the client code is actually quite straightforward. So,
for the time being, this package will only provide the mat-vec-3 functor.